Part 4-d exercises

+++ Part-4/blog_list/controllers/blogs.js
@@ -1,11 +1,11 @@
 const app = require('express').Router();
-const { request } = require('../app');
 const Blog = require('../models/blog');
+const { userExtractor } = require('../utils/middleware');
 
+// Fetch all blogs
 app.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({});
-    console.log("Fetched blogs:", blogs); 
+    const blogs = await Blog.find({}).populate('user');
     response.json(blogs);
   } catch (error) {
     console.error("Error fetching blogs:", error.message);  
@@ -13,19 +13,26 @@ app.get('/', async (request, response, next) => {
   }
 });
 
-app.post('/', async (request, response, next) => {
+// Create a new blog
+app.post('/',userExtractor, async (request, response, next) => {
+  const { title, author, url, likes } = request.body;
   try {
-    if (!request.body.title || !request.body.url) {
-      return response.status(400).json({ error: 'Title and URL are required' });
+    const user = request.user;
+    if (!user) {
+      return response.status(401).json({ error: 'Unauthorized' });
     }
     const blog = new Blog({
-      title: request.body.title,
-      author: request.body.author,
-      url: request.body.url,
-      likes: request.body.likes || 0  
+      title,
+      author,
+      url,
+      likes: likes || 0,
+      user: user._id
     });
 
     const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog.id);
+    await user.save();
+
     response.status(201).json(savedBlog);
   } catch (error) {
     console.error("Error saving blog:", error.message);
@@ -33,22 +40,31 @@ app.post('/', async (request, response, next) => {
   }
 });
 
-app.delete('/:id', async (request, response, next) => {
+// Delete a blog
+app.delete('/:id',userExtractor, async (request, response, next) => {
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(request.params.id);
-    if (!deletedBlog) {
+    const user = request.user;
+    if (!user) {
+      return response.status(401).json({ error: 'Unauthorized' });
+    }
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
       return response.status(404).json({ error: 'Blog not found' });
     }
+    if (blog.user.toString() !== user.id.toString()) {
+      return response.status(403).json({ error: 'Unauthorized to delete this blog' });
+    }
+    await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
   } catch (error) {
     next(error);
   }
 });
 
+// Update blog likes
 app.put('/:id', async (request, response, next) => {
   try {
     const { likes } = request.body;
-
     if (likes === undefined) {
       return response.status(400).json({ error: 'Likes field is required' });
     }
@@ -69,6 +85,5 @@ app.put('/:id', async (request, response, next) => {
     next(error);
   }
 });
-
 
 module.exports = app;
